+++ admin/Model/StudentSubject.php
@@ -107,9 +107,17 @@ class StudentSubject extends AppModel {
 	// 1.studentsubject delete where student_id = $id
 	public function delete_student($id)
 	{
-		if ($this->deleteAll(['student_id' => $id]) == false)
+		$find_student = $this->find('all', [
+			'conditions' => [
+				'student_id' => $id,
+			],
+		]);
+		if (!empty($find_student))
 		{
-			return false;
+			if ($this->deleteAll(['student_id' => $id]) == false)
+			{
+				return false;
+			}
 		}
 		return true;
 	}
